+++ @utahdts/utah-design-system-header/artifacts/index.d.ts
@@ -8,7 +8,7 @@ declare module "@utahdts/utah-design-system-header" {
   export type Environments = "none" | "a1" | "a2" | "a3" | "custom" | "unittest";
   export type Events = "utahHeaderLoaded" | "utahHeaderUnloaded";
   export type HeaderApplicationTypes = 'wordpress' | 'static site' | 'salesforce' | 'custom application' | 'microsoft power apps' | 'servicenow';
-  export type PopupPlacement = "auto" | "auto-start" | "auto-end" | "bottom" | "bottom-start" | "bottom-end" | "left" | "left-start" | "left-end" | "right" | "right-start" | "right-end" | "top" | "top-start" | "top-end";
+  export type PopupPlacement = "bottom" | "bottom-start" | "bottom-end" | "left" | "left-start" | "left-end" | "right" | "right-start" | "right-end" | "top" | "top-start" | "top-end";
   export type Size = "SMALL" | "MEDIUM" | "LARGE";
   export type UtahIdFetchStyle = "Automatic" | "None" | "Provided";
   export type MainMenuItem = {