+++ utah-design-system-website/src/react/components/websiteContent/HomeLanding.jsx
@@ -1,5 +1,5 @@
 /* eslint-disable max-len */
-import { Accordion, ExternalLink, MainContent } from '@utahdts/utah-design-system';
+import { Accordion, ExternalLink } from '@utahdts/utah-design-system';
 import { Link } from 'react-router-dom';
 import layoutScreenshot from '../../../static/images/screenshots/layout.webp';
 import { LightBox } from '../lightbox/LightBox';
@@ -9,7 +9,7 @@ import { WcagRulesBlock } from './guidelines/WcagRulesBlock';
 
 export function HomeLanding() {
   return (
-    <MainContent className="landing-page-template">
+    <section className="landing-page-template">
       <div className="home-banner">
         <div className="home-banner__title">Utah<br />Design<br />System</div>
       </div>
@@ -276,12 +276,12 @@ export function HomeLanding() {
           <IconsWebsite.IconChatBubbles isHidden />
           <h2 className="text-center mt-spacing-l">Don&apos;t be Shy!</h2>
           <p>
-            Have a question? <a href="mailto:dts_ui@utah.gov">Contact us</a>. Got a suggestion? <a href="mailto:dts_ui@utah.gov">Contact us</a>.<br />
+            Have a question? <a href="mailto:dxp@utah.gov">Contact us</a>. Got a suggestion? <a href="mailto:dxp@utah.gov">Contact us</a>.<br />
             We like hearing from you, and this space is for all of us; Help us make it better!
           </p>
         </div>
       </div>
 
-    </MainContent>
+    </section>
   );
 }