+++ utah-design-system-website/src/react/components/websiteContent/guidelines/AccessibilityDocumentation.jsx
@@ -30,6 +30,8 @@ import { PreCodeForCodeString } from '../../preCode/PreCodeForCodeString';
 import { pageUrls } from '../../routing/pageUrls';
 import { StaticExample } from '../../staticExamples/StaticExample';
 import { IconsWebsite } from '../IconsWebsite';
+import sizeExample from '../../../../static/images/screenshots/spacing/size_target.webp';
+import spacingExample from '../../../../static/images/screenshots/spacing/size_spacing.webp';
 
 export function AccessibilityDocumentation() {
   const goodAltAudioVtt = new URL('../../../../static/audio/vtt/GoodAlt.vtt', import.meta.url).href;
@@ -190,6 +192,13 @@ export function AccessibilityDocumentation() {
               The link text should accurately describe where or what is being linked to. This will also help your page achieve a better SEO score.
             </li>
             <li>
+              Links should be easily identified from the rest of the content on a page.
+              Use both color and some form of styling (underlined, bolded, or italicized) to highlight them (links as part of navigation are exempt).
+            </li>
+            <li>
+              Links should have a clear hover state: underline or background color change.
+            </li>
+            <li>
               Links that open a new tab have a visual indicator and visually hidden text like “opens in a new tab” to warn screen reader users.
               For example: <ExternalLink href="https://www.google.com/search?q=arches+national+park">Arches National Park</ExternalLink>
             </li>
@@ -201,7 +210,7 @@ export function AccessibilityDocumentation() {
             <li>
               Links to email addresses should include the actual email address in the text of the link.
               Additionally, if there is no text to indicate the link is for an email address, include visually hidden text to inform screen reader users.
-              For example: <a href="mailto:dts_ui@utah.gov"><span className="visually-hidden">Digital experience team email:</span>dts_ui@utah.gov</a>
+              For example: <a href="mailto:dxp@utah.gov"><span className="visually-hidden">Digital experience team email:</span>dxp@utah.gov</a>
             </li>
           </ul>
         </li>
@@ -538,7 +547,7 @@ export function AccessibilityDocumentation() {
           is critical to understanding, be sure to give them an alternative source for the information.
         </li>
         <li>
-          <strong>Interactive content.</strong>  When possible, ensure that the user:
+          <strong>Interactive content.</strong> When possible, ensure that the user:
           <ul>
             <li>Can activate or dismiss content using the keyboard.
               <ul>
@@ -564,9 +573,26 @@ export function AccessibilityDocumentation() {
                 </li>
               </ul>
             </li>
-            <li>Will be able to easily click the object (has a large clickable area).</li>
             <li>Has enough time to complete necessary tasks.</li>
             <li>Can correct any errors that may have been made.</li>
+            <li>
+              Will be able to easily interact with controls.
+              Interactive elements should have a size of at least <strong>24 by 24 pixels</strong>, without overlapping with other elements.
+              <div className="text-center"><img src={sizeExample} alt="Example of acceptable and unacceptable size for interactive elements" style={{ maxWidth: '100%' }} /></div>
+              <ul>
+                <li>
+                  <strong>Exception.</strong>
+                  <ul>
+                    <li>When the element is positioned inline, as a part of some text. Rely on the line-height instead.</li>
+                    <li>When the element is deemed essential. For example, locations on a map that are close together.</li>
+                  </ul>
+                </li>
+              </ul>
+              <ul>
+                <ul><li>If elements do not meet the minimum requirement, you can improve the user experience by accounting for the <Link to={pageUrls.spacing}>spacing</Link>.</li></ul>
+              </ul>
+              <div className="text-center"><img src={spacingExample} alt="Example of acceptable and unacceptable spacing between interactive elements" style={{ maxWidth: '50%' }} /></div>
+            </li>
           </ul>
         </li>
       </ul>
@@ -586,9 +612,8 @@ export function AccessibilityDocumentation() {
           and adjust the volume.
         </li>
         <li>
-          <strong>Captions.</strong> Captions provide a text based alternative that is synchronized to the audio being played. It is highly
-          recommended to use a media player that also allows the user to change the size and color of the captions. These can also be
-          known as &quot;sub-titles&quot; and is available with most media players. To view more information on captions see
+          <strong>Captions.</strong> Captions provide a text based alternative that is synchronized to the audio being played. To ensure optimal accessibility, use a media player that also allows the user to change the size and color of the captions. These can also be
+          known as &quot;sub-titles&quot; and is available with most media players. To view more information on this requirement see
           the <ExternalLink href="https://www.w3.org/WAI/media/av/captions/">w3.org page on Captions/Subtitles</ExternalLink>.
           <StaticExample
             renderedExample={<LightBox image={captionsImg} alt="Video providing captions" />}
@@ -596,9 +621,9 @@ export function AccessibilityDocumentation() {
           />
         </li>
         <li>
-          <strong>Transcripts.</strong> Transcripts refers to a written version of the spoken content in an audio or video file. Not only
+          <strong>Transcripts.</strong> Transcripts refer to a written version of the spoken content in an audio or video file. Not only
           does this benefit individuals who are deaf or hard of hearing, but also people who may have difficulty understanding or processing
-          spoken language. For more information on how to get or create transcripts please refer to
+          spoken language. For more information on this requirement and on how to get or create transcripts, please refer to
           the <ExternalLink href="https://www.w3.org/WAI/media/av/transcripts/#creating-transcripts+">w3.orgs&apos; information on Creating Transcripts</ExternalLink>.
         </li>
       </ul>